Fix missing broadcast ANR culprit information.
All TimeoutRecord have unfortunately been marked as "unknown culprit"
due to the Intent not having a ComponentName filled in.  We fix this
by passing the relevant package and class name through the record,
and we expand it to identify runtime registered receiver classes to
aid faster bug triage.

For now, we're abusing the class name transported via "receiverId"
to avoid the churn of adding yet another method argument late in the
release cycle.

Bug: 266169149
Test: atest InternalTests:TimeoutRecordTest
Test: atest FrameworksMockingServicesTests:BroadcastQueueTest
Test: atest FrameworksMockingServicesTests:BroadcastQueueModernImplTest
Test: atest FrameworksMockingServicesTests:BroadcastRecordTest
Change-Id: I67b6f672c24627b7281de3b218e1114ab8a48f74

Former-commit-id: a8c29399f8c01fe6a04b7f0b1d2f5934ae7de222