Don't throw if the matching key contains ":".
If a client adds custom crafted strings as part of matching key,
it is possible for this restriction to go unnoticed. Instead,
just all allow characters in the key. This means that we cannot
solely rely on getDeliveryGroupMatchingKey() for matching but
don't think this would be a problem for clients of this API
outside the System.

Bug: 275312433
Test: TH
Change-Id: I2ad9c6d270678ddb881168936b1f74bd375de58c
Merged-In: I76e307800017d4dd1e66233b1ff4cc5cbcef7246

Former-commit-id: a72e432a0fecac79b562ac5c11f0629549ce0bcc