Move implementation details out of VDM
Add a new VirtualDeviceInternal class, which holds all the internal
implementation details of VirtualDevice's interaction with other
services. All its methods are package private and do not contain
unnecessary annotations and documentation.

VirtualDeviceManager and VirtualDeviceManager.VirtualDevice now are
the single go-to point for API and documentation, without unnecessary
clutter.

Fix: 261163570
Test: atest CtsVirtualDevicesTestCases
Change-Id: Iee610e162744d0cafd0925b3e3492bb4dd273403

Former-commit-id: eda545b0eaf16ca9f3904440a2c2721b242a281d