Define FeatureFlagsService.
This service acts as a central coordinator to ensure that all
cross-process, stable flags maintain the same value, as well as
providing a central mechanism for overriding flags locally on a
device.

This cl contains a the service wholly defined and functional, though
the FeatureFlags class in core, meant to communicate with it, is not
yet implemented.

See http://go/android-flagging-dd for more details.

Test: build && flash && `adb shell cmd flags`.
Bug: 279054964
Change-Id: I930152682f98f3f299e763522e5c7e6b1a550e32

Former-commit-id: 4f7ccd42f17376e05e37e7c0499171b3542330f1