Pausing the transient hide activity with user leaving hint
Activity#onUserLeaveHint was not called while swiping up to home
since the activity was directly added to stopping list when
transient launch transition completed.

The code added in commit 0572aac to send task-stack-changed
event while finishing transition was also updated. It is not
needed if the activity is scheduled to be paused by TaskFragment
#startPausing since the task-stack changed is called after pause
completed. However, it is still needed if the activity enters
auto-pip when transition finishes.

Bug: 274576791
Test: swipe up to home
Change-Id: Idf664a93e95471d83c0964dc67ab384521f78b24

Former-commit-id: 4dd7aefbb525015066192988ebdfaa980a95999c