Prevent flicker on orientation change while starting
It's common for apps to set the orentation in the on fragment attach or
some otherway that delays that orentation change to a point when the
opening activity transition has alrady been started. In this case the
orientation change will trigger an activity restart but the position of
the activity and letterboxing are updated before the activity re-draws
itself with the new configuration, which leads to a flicker.

Test: use test app from bug and make sure issue doesn't reproduce
Bug: 271276745

Change-Id: I303721bb33bbca32e2e65b7c6f1e0905771d221a

Former-commit-id: 7d82f0a2b83a041d49780f61125c32725552521a