[hotword] init audioflinger outside of lock
The lock used in HotwordDetectionConnection is shared with VoiceInteractionManagerServiceStub, meaning UI thread can end up waiting for audio flinger service to initialize, causing a freeze.

Not that this is not a perfect fix, since the first call (in HotwordDetectionConnection constructor) is still locked. However, the deadlock seems to occur when the service dies and we try to re-initialize, which this change should fix.

Bug: 215887032

Test: atest CtsVoiceInteractionTestCases

Change-Id: I35c682910ac6605041175ab523a71f2b83b499dc

Former-commit-id: a1eb2797534b4b1ac35baef95c1f2ebf2d08d739