Fix transition timeout when launching activity in TaskFragment
Transition timeout due to the sync group was not ready. It was
set to false when an activity was started while a transition
was already collecting. Somehow, the activity was resumed
already because there was no need to pause the previous top
activity. So, the transition won't be set to ready afterwards.

Bug: 274540105
Test: verified both on sample app and Settings
Change-Id: Ib18652f610b448beb12eeb98a473399832d71dc5

Former-commit-id: c058b8aa5b41aca47d18d19d49ff81a1ebbd2ba3