Remove receiver flag enforcement workaround for affected package
Android U enforces the requirement that an app registering for
non-system broadcasts must specify a flag indicating the exported
state of the receiver. During the initial enforcement period, some
apps were not yet ready for this enforcement, so a workaround was
added to the platform to allow time for these apps to flag their
affected receivers. Now that all affected apps have been updated,
this commit will remove the workaround to ensure the enforcement
is applied to all apps targeting U+.

Fixes: 259139792
Test: Manually verified affected apps were no longer reported
Change-Id: I1ce044894d6d2badca8aaa9b87d4a5c53f895bc5

Former-commit-id: 2bf1f4d0bdd97fbbdde3af94c41a16c765e6a5fe