Disabled SUBPIXEL_TEXT_FLAG for notification shade while fold unfold
Disabled the SUBPIXEL_TEXT_FLAG for all TextViews of the Notification Shade while folding and unfolding is in progress.

The SUBPIXEL_TEXT_FLAG is responsible for the default Android behaviour that tries to align the font to the nearest pixel, causing the text to jump as the nearest pixel keeps changing during folding and unfolding.

To solve this issue, we have disabled the SUBPIXEL_TEXT_FLAG for the TextViews of the Notification Shade during folding and unfolding, resulting in smoother text transitions.

Test: atest com.android.systemui.shared.animation.DisableSubpixelTextTransitionListenerTest.kt
Fixes: 227427854
Flag: SPLIT_SHADE_SUBPIXEL_OPTIMIZATION
Change-Id: I6ada51f912ee1c086674b19ea39d4cab3ad161a8

Former-commit-id: 32de1c9474dd22ff9bc43d13643bfeb7bbc20fee