Ensure scrim visibility updated by mPrimaryBouncerToGoneTransition
This was a race condition where the back scrim was not getting to
alpha=zero in time for the last scrim visibility update, so it was
keeping the window up (which then ate the touches). The fix is to
add an insurance scrim visibility update at the end of the fade-in.

Test: ran all atest tests
Test: manually verified that the breaking condition occurred but was mitigated
Fixes: 273998188
Change-Id: I2986c82b3c6831fc581e745cd2a3ac118f314f33

Former-commit-id: ceb7010da6a5c0301b506ef43da93ba785f52901