Fix tap outside focus for SCVH from different process.
When granting embedded window focus due to a tap, send a null
WindowSession. This is because it's not necessary for the host and the
embedded to be in the same process. So when WMS looks up the host
window, it doesn't need to check that the embedded session matches the
hsot session. WMS is already providing the correct host window token for
the embedded window so the additional check is not needed.

Test: SurfaceControlViewHostTests
Bug: 230340812
Change-Id: I9bb11a4d5283ee3e3fa5ac9972630dbaefff921f

Former-commit-id: ec7100fe88a7381f1180c4914d06cdac3778bbb6