Fix media showing AOD in split shade and double-line clock disabled
For split shade, media is not automatically hidden when going to AOD.
When using the double line clock, this is not an issue, as the clock
container takes the whole size of the keyguard status container, hiding
also media.
For single line clock, the clock container uses less space, making media
visible.

In single shade, media is part of the NSSL container, which already gets
hidden during AOD, also hiding media.

Test: KeyguardMediaControllerTest.kt
Test: Manually - Single shade, and split shade, with
                 the double line clock setting enabled and disabled,
                 while media is showing. Then going from/to AOD.
Fixes: 276682118
Change-Id: Ie2f4ac3aab4bcdb69f33c57292ff0c9a073bbcae

Former-commit-id: 66d35542299ddfdbb4ab30b54f318c023652e94a