End animation when deleting.
It seems that when we delete and add super quickly, the delete method is
preventing the view from being removed. Instead of removing the view
after the full duration of the animation is complete, we end the
animation to remove the view to begin the next delete.

This actually also makes the reset animation much more succinct when we
long press delete.

Fixes: 276400809
Test: add and delete in pin view very quickly
Change-Id: Ica145faa95a342b10832e355fe04201a9fc78459

Former-commit-id: 2797d459142ec6d237ca3365ee46d559ca5cdda3