Replace NSSLController usage in NotifShelf w/ NSSL
Under a unidirectional architecture, Views should not have access to Controllers.

In the future, we may want to explore decoupling NotificationShelf from NotificationStackScrollLayout entirely, and wire up their interaction externally via ViewBinders / Controllers. That is out of scope of this CL.

Note that we had to pull in NotificationRoundnessManager as a dependency in NotificationShelf. This is needed for functionality that was originally exposed via the NSSLController. NRM is effectively a View class, serving as a Data Store for the NSSL hierarchy, so it's OK w/r/t unidirectionality, especially considering NotificationShelf has a reference to the NSSL anyway. In the future, when we move some of the shelf logic back into the NSSL / StackScrollAlgorithm (specifically, the part that traverses the children of the NSSL), then we can drop the dependency on the RoundnessManager.

Bug: 271161129
Test: atest SystemUITests
Change-Id: I5887c0b0c62d2464f8818fdf42f8e38484f3a59a

Former-commit-id: 6e20c37aeb9fc272a8c6e1267ddf6b1962beae92