Unify USE_FULL_SCREEN_INTENT permission check.
So that the developer API and the service enforcement will always be
consistent.

Also fixes the bug where NotificationManager.useFullScreenIntent()
isn't taking flagging and pre-Q target SDK into account.

This changed FSI_FORCE_DEMOTE to work for pre-Q apps as well, since
it's a development flag and making it work for all apps is more
consistent.

Bug: 277104139
Test: NotificationManagerTest
Test: NotificationManagerServiceTest
Change-Id: Ic5736ef2c9ad239dbe822147641ce5db5436dd53

Former-commit-id: 1fb6b61687f756bc62c7786a3e37ba23c1265122