Make sure topActivity always retrieves focus when move task to back
In a freeform environment, multiple windows can be visible and
therefore RESUMED at the same time. It means that when moving a task to
back, the topActivity that is going to be focused may be RESUMED.
However, there's a check in `moveTaskToBackInner()` to only trigger
`resumeFocusedTasksTopActivities()` when topActivity is not RESUMED,
which results in DisplayContent#mFocusedApp not getting updated
correctly.
This CL fixes the bug by triggering DisplayContent#setFocusedApp() even
topActivity is RESUMED.

Bug: 276136272
Test: Manual testing to make sure the top activity can get focused
Change-Id: I51371394830277777ff7983fd1d1c14d63fa505a

Former-commit-id: 004e857ce5cb60b3bfab8756bdf5df1779f11924