Report visibility only when display is fully on in WallpaperService
This is needed as some custom wallpaper interactions need to start when the screen turns on (which is not the same as "wallpaper visibility" or "wake" status).

Previously, only the power state was taken into consideration. However, brightness is turned on after all windows are drawn (brightness is separated from power state).

This was causing some wallpapers to trigger animations while the screen was not yet visible.

Bug: 275057152
Test: Recorded perfetto trace and made sure visibility callback is invoked at the correct moment (after display brightness is turned on)
Change-Id: I8c8c59ad0e3f19880ec543686285a89b0b864739

Former-commit-id: 42e1f4f58a5b4fc3cfdd27c7a02e5b53e954aae0