Fixes RTL bug in quick affordance slot tabs.
When rendering with a right-to-left language, we shouldn't flip the
order of the slot names; this way, the "right" slot tab remains on the
right and the "left" slot tab remains on the left, regardless of the
layout direction of the current locale.

Fix: 276477249
Test: included unit test
Test: manually verified that "Left shortcut" is always on the left and
"Right shortcut" is always on the right, regardless of the direction of
the current system language (used Hebrew to verify).

Change-Id: Ibcffff3b4f26231c78299fd8afe4a16fe5cb1b11

Former-commit-id: 51f86a1592e78498e76efbead5841059c2698be4