Fix system crash due to GREF overflow
The TCMS uses LruCache to save sessions, it will handle unlink
DeathRecipient when session is destroyed. But the TCMS doesn’t
handle unlink DeathRecipient if there are many sessions that exceed
the cache size.

In this change, the system will unlink DeathRecipient if the session
is evicted from LruCache. But this change doesn't handle the behavior
if the elements are evicted from the cache. This is unrelated to the
crash, it will be handled later.

Bug: 275298870
Test: local update cache size to 1, use testMultipleActiveSessions()
to verify the evicted behavior.

Change-Id: Ie0f23b26005c73080473a1550ef7d1f275bf7e27

Former-commit-id: 4621a0c928a26004191b79d3bd4dccf7f1e1cccb