Run update note task shortcut in a service if user is different
Shortcut Manager injects the main user, and do not allow us to set what user should have their shortcuts modified. To workaround that until we add support to custom users, we are launching a service and triggering the update task from that user's process.

Running a different service may delay the update by a fraction of a second, in case the user jump into the home screen just after.

Test: atest NoteTaskControllerTest

Fixes: b/276889922
Change-Id: If503df7af038c51829e9267af7a9d0aafd7da986

Former-commit-id: c2beb67c03a8a0a69ef92a741ff43ea4b91d0038