Update OWNERS for internal Chooser and Resolver
This change reduces OWNERS for the relevant files down from
all of SysUI to just the team actively maintaining it. It also
adds the correct bug component where necessary.

Background:

ChooserActivity development has moved to a new project:

  *  packages/modules/IntentResolver

The framework version of ResolverActivity is presently
still in use at the time of this change, but this will also
move in the near future. Most changes should be made
in IntentResolver first, then backported here if necessary,
depending on the situation.

Bug: 277242209
Test: presubmit linter
Change-Id: Ia4271556b7f705bbf1a72d35433622e917da5dd4

Former-commit-id: b7adc1ba146fa00668240de984bca9d51dd83cea