Update positioning during drag using SurfaceControl.Transaction
When the bounds are updated pixel by pixel using a WCT in task
positioner, it updates mLastNonFullscreenBounds. So, when the task is
dragged to the top to transition into fullscreen,
mLastNonFullscreenBounds is updated, and, when we transition back to
freeform using the handle menu, the positioning of the task is at the
top. By using SurfaceControl.Transactions to update the positioning,
mLastNonFullscreenBounds is left untouched when transitioning to
fullscreen via dragging.

Bug: 277137896
Test: Drag task from freeform to fullscreen and then transition task
back to freeform using handle menu. Task should be in positiom before
drag.

Change-Id: Idba8523e3c8851b30ef5938b921e9b4ed23bcefa

Former-commit-id: 8aa9d4d288b58bfc72b9eda56ebe81c1a3dae5bc