Long-press lock screen popup support for large font size.
Changes the height of the popup from a fixed height that matches the
lock screen shortcut button heights to a wrap_content to allow the popup
to be taller if it needs to accommodate biggfer fonts.

Also moves things around a bit to make sure the lock screen shortcuts
and the popup are bottom aligned with each other, simplifying some of
the layout properties as a result.

Fix: 277279440
Test: manually tested the popup with both the biggest font size and the
default font size. Made sure that the popup is always bottom aligned
with the lock screen shortcuts and the text is properly seen (even if
it's ellipsized).

Change-Id: I924a15609d52c09ec028df1f7f62b6b9efeab1fe

Former-commit-id: 8d67255b57adf6d6eb4181b4897b816304953414