Block relayout during transition to desktop
After starting a transition, DesktopModeWindowDecorViewModel#onTaskInfoChanged runs before the transition animation. This runs relayout on the task which causes a flashing affect during the transition since relayout attempts to update the task surface to the post-transition state. This change allows us to block task surface changes in relayout until the transition is finished or aborted.

Bug: 277941322
Test: Transition to freeform by dragging fullscreen task down from caption handle. When released, task should not flash.
Change-Id: I0242d7765685bfc692739915263f1b634db16da4

Former-commit-id: 1ed33484360df9c772ede2566a56513537c29063