Only override embedded activity locale if source package lookup yields locales to match
Calling Configuration#setLocales also sets the text direction for the containing Activity. If we pass an empty LocaleList to this method, then we lose the intended text direction for its locale. So this change will only override the target (embedded) Activity's locales if our package lookup for the host yields a populated LocaleList to provide.

Bug: 277363604
Test: Verified on Pixel devices in Settings app

Change-Id: I8bcdab0737c18f861f74a8f9bd2d15043359e27b

Former-commit-id: d9bc801301acdf24fda7ff043da35f70631948b1