Allow Recents transitions to collect in parallel
We can't fully support parallel transition collection,
but we can for specific cases (like recents). This CL
adds some hard-coded checks for recents specifically and
allows it to collect in parallel with activity-only
transitions. This is basically way of assigning "priority"
to recents since it is a very user-forward operation
and an integral part of how users percieve device
performance.

Bug: 277838915
Bug: 264536014
Test: In the running task, launch a slow-to-draw activity,
      then swipe to recents or home. It should not wait
      anymore.
Change-Id: Ie09e31bf290847b108578b688f57360536fe7f4e

Former-commit-id: b4963b6da5669dc62b37a0e7ed80091e43203b73