BroadcastQueue: Initial deterministic perf test.
As we start working on the "modern" alternative broadcast queue
implementation, we'd like to gain confidence that we're actually
improving the metrics we care about, without having to wait for
changes to deploy to wider user populations.

Instead of writing these using a traditional timing approach (which
can suffer from scheduling noise), we explore "playing back" a set
of well-known pathological scenarios, which will result in a
deterministic output that can be used to immediately judge if a
given CL improves or regresses a measured metric.

Bug: 245771249
Test: atest FrameworksMockingServicesTests:BroadcastQueueTest
Change-Id: I6f562c0b6f2a71fef52aae013899de7f2f95bc71

Former-commit-id: 3f400959e29ba57ed69676f87744fb887422ef30