Skip stopping the BG services for the new processes
Processes are created as UID idle state, and we may emit
the UID IDLE event after it's being created. However this would
stop the pending bg services if they happen to be in the middle
of restart, which doesn't make sense here. So we don't stop
the bg services if its a new process here.

Bug: 272375589
Test: tree hugger
Change-Id: I1a7f2c963295b2f42c19ddf73dd59583d65ba02f

Former-commit-id: 6dc489c8ffdf816245fd1bd8c3410b478169d3ea