Re-show faded PiP when fixed rotation is finished
PiP will be faded out if there is orientation change when
switching app (onFixedRotationStarted), and then fade in with
display transition from startAnimation -> fadeExistingPip.

But if there is no app switch, such as entering recent and
then return to current app, the signal to re-show the PiP was
missed in shell transition. So still listen to
onFixedRotationFinished and restore the visibility of PiP
when the transition is idle.

Bug: 262835608
Test: Enter PiP in portrait. Launch a rotatable app.
      Rotate screen to landscape.
      Either:
       - Swipe up and down to cancel.
       - Swipe up to enter recent.
         Click the current app to return.
      Though the PiP will be fade out during swipe-up gesture.
      After return to current app, the PiP should be still visible.

Change-Id: Ieb47eceaa5e8e7ff710a0fd6ab8319d1a3463985

Former-commit-id: bf1818439f5c0e509a1e5b4ac87c0b4d091a2a5a