Adapt the boot logic for lockscreen live wallpapers
First, if there are two live wallpapers, it is necessary to call bindWallpaperComponentLocked twice get the proper lockscreen live wallpaper after boot or after a user change switch.

We only call bindWallpaper, we do not create a second WallpaperObserver:
the observer from the system WallpaperData takes care of listening to
everything. Also, we only use one callback, not two.

Second, we need to handle lockscreen live wallpapers that are not
direct-boot aware. If that is, we need to show the default wallpaper
until the device is unlocked. In order to do that, we need to delete the
LOCK_ORIG and LOCK_CROP files, to avoid showing an outdated wallpaper.

Bug: 253507223
Test: Manually with the LWP feature flag enabled,
  -set a NOT direct-boot aware wallpaper on home + lock screen,
  -set another live wallpaper on home screen to trigger the migration,
  -set a pin code on the device
  -reboot
  -do the same for different scenarios (two NOT direct-boot aware, or a
  not direct-boot aware wallpaper on home screen only)
Test: atest CtsWallpaperTestCases

Change-Id: I353359a90e4bddc6dcdde88c9a7ab5a616a68bd3

Former-commit-id: 46c361b28b665878afa4d5e29c58f62d7802e573