Separated animation of status bar in split shade
This change fixes the bug where date view from status bar of notification split shade is not animated while folding and unfolding is in progress.

The root cause is that the incorrect date view is being animated due to multiple views with the same ID under the same parent ViewGroup(Split shade).

Hence we separated the animation of the status bar of the split shade from other views in the split shade.
This allowed us to use the split shade status bar as the parent ViewGroup to correctly register the date view for animation.

Test: atest com.android.systemui.shade.NotificationPanelUnfoldAnimationControllerTest
Test: functional test in a follow-up cl
Fixes: 277214741
Change-Id: I27d33721b1c835a51587f3865898a1e8a48f3f3c

Former-commit-id: 3c18f4418d9b4ebee8e6a749571a32a5df6ae83d