Maintain keyboard configuration and reload layouts if changed
Virtual keyboard creation logic involves 2 steps:
- Create a uinput device
- Updating associated layout info
It is possible that when onInputDeviceAdded() occurs layout info is
not correctly set. So, everytime we get onInputDeviceChanged()
callback and onImeInfoChanged() callback there is a possibility that
keyboard layout needs to be reconfigured. So, keep track of current
configuration and call reload() when a new layouts needs to be set.

Test: atest VirtualKeyboardLayoutTest
Bug: 277778640
Change-Id: I998009d394a2103bf5afb37a622094dd771162d8

Former-commit-id: eaed646cb9c86e283ccb1cfa4998fcb8796a3564