Fix allow-UIJ-scheduling determination.
The ServiceRecord.fgRequired bit is updated after the call to check the
FGS while-in-use permission. This meant that trying to set the
allow-uij-scheduling bit in that method while limiting it to foreground
services didn't work properly. Instead, wait until the system has
confirmed the FGS can start and updated the ServiceRecord bits before
setting the allow-uij-scheduling bit.

Bug: 277635240
Test: atest CtsJobSchedulerTestCases:UserInitiatedJobTest
Change-Id: Iccb8a9b2ee39745f8f9c9d5db4bcf969cf175997

Former-commit-id: 6b3130423780176785071d4f766d4010026256b6