Make snooze view taller
When notification guts are shown, we don't make the backing notification
card invisible, we just clip it to the height of the guts view.

However, in S we introduces pretty severe corner rounding, which means
that part of the clipped NotificationContentView can still be visible
behind the guts view.

So make the view a bit taller to account for the corner rounding

Fixes: 191935939
Test: manually view snooze view for each template type
Change-Id: Id5fa910646f78349199d1f47ae7724b8b15e764a

Former-commit-id: 45e962ff2a84420091b84c232487292d0c0f25db