Fix location provider mismatch.
Requesting location updates without specifying a provider tended to use
the FUSED provider under the hood, but DeviceIdleController was checking
for the NETWORK provider instead. Switch to explicitly check for the
FUSED provider and request updates from the FUSED provider so that the
system can be consistent and avoid system crashes if the FUSED provider
is missing.

Bug: 274369433
Bug: 274572355
Test: atest DeviceIdleTest
Test: atest FrameworksMockingServicesTests:DeviceIdleControllerTest
Change-Id: I98bab13d3cc685f504c934b4714543d1b16a612c

Former-commit-id: b5c80397848b60f0da576552a530ec998fd35ed0