SlicePurchaseActivity don't send canceled on manual finish
onDestroy is called when the user closes the app or when finish() is
called programatically. We call finish after sending a response from the
slice purchase activity, but we shouldn't send the canceled response
because the user didn't close the application. Add a check to see
whether onDestroy was called programatically or due to user action.

Test: manual, CTS
Bug: 276703840
Change-Id: Iae1f5e77a1b6d75b911f4bd5314869b36157afda

Former-commit-id: 1e26c2445bcc16f3da94649fa0eead32678e5bac