Fix NPE when setting icon overlay in PipTaskOrganizer
Similarly to check in PipTransition, we need to add similar condition
in PipTaskOrganizer which is used when shell transitions are disabled.

If the top activity info is missing, we fallback to color overlay.

Bug: 277637354
Bug: 272819817
Test: n/a, this is addressing a bug in field, the investigation why
this happens will be continued in 272819817, and it simply copies the
same condition that was used in PipTransition (as the two codepaths are
similar in that sense) in http://ag/21965862

Change-Id: Iab20d3b53026757d1bf3f8714aa750b8cd8c2c40

Former-commit-id: 896a91b764150e112a5395323f2deb79f81fa67d