Fix how we send flag changes to mDwpcHelper in relayoutWindow
We were accidentally conflating the exclusive OR bitmask of which
flags have changed, and the flag values themselves. This was causing
two problems:

(1) if a flag like FLAG_SECURE is added early in an Activity lifecycle
such as in onCreate, the first layout happens with the flag present
and therefore it wouldn't be included in the exclusive OR bitmask
value sent to DisplayWindowPolicyController (which was expecting the
flag value itself).

(2) If a flag happened to have been removed, it would end up in the
exclusive OR bitmask and get passed to DisplayWindowPolicyController,
which would think that flag is present, not removed.

Fixes: 261991662
Test: atest WmTests:WindowManagerServiceTests
Change-Id: I18c359665264e689fcecea6428e3b45dc6337983

Former-commit-id: 5db377ff53b84a2f55364bb4edfed8591a428463