Add logging for zen mode events.
Gated on the LOG_DND_STATE_EVENTS flag.

This change adds a ZenModeEventLogger that logs atoms on zen mode state changes, policy changes, and changes in the number of active automatic rules. Additionally adjusts some methods in ZenModeHelper to pass the calling UID through (not gated by flag), which is only used when logging.

Bug: 259261349
Test: ZenModeHelperTest, PreferencesHelperTest, NotificationManagerServiceTest, AudioManagerTest (cts), NotificationManagerTest (cts), CtsLegacyNotification27TestCases, manual via statsd_testdrive
Change-Id: I4ed2d664b501b3cf281aa70a90a1d4fb8cebdf2e

Former-commit-id: 7e1156b658014e3f52ccdacb93ebfd193e6340fd