Defer remove splash screen window until start transaction apply.
In previous CL ag/22494388, there try to defer remove starting window
if it is in collect, which miss the case that transition may already
playing.
So in this CL, by override the waitForSyncTransactionCommit and
onSyncTransactionCommitted, there can defer remove starting window
until the starting transaction apply, which should ensure the surface
of app window is drawn.

Bug: 276692049
Test: manual launch app from notification shade several times, verify
no blank window occur.

Change-Id: I4ad3f4c892939ee70cbb544d63be03123c7425bc

Former-commit-id: d7fb5d979ce2adc7d33be14548224a83d471b0b2