Implement soundbar volume UI for TV panels.
This is done by adopting adjust-only absolute volume behavior when the
connected amplifier is able to send <Report Audio Status>. Adjust-only
absolute volume behavior supports volume tracking and automatically
displays volume UI when the volume level is updated in AudioService.

In HdmiControlService, adoption of this behavior is implemented as a
special case in the logic for adopting absolute volume behavior:
- Absolute volume behavior is currently adopted when a TV panel or
  playback device is able to track AND set the percentage volume level
  of the System Audio device.
- Adjust-only absolute volume behavior will be adopted when a TV panel
  can track, but not set, the volume level of the System Audio device.

The only behavioral difference in HdmiControlService when using
adjust-only AVB, compared to regular AVB, is that it does not send <Set
Audio Volume Level> to the System Audio device.

Bug: 240663266
Test: atest TvToAudioSystemAvbTest PlaybackDeviceToTvAvbTest PlaybackDeviceToAudioSystemAvbTest
Change-Id: I05465cd0eab2f3c3464bf41c1194cdc3be7af4bf

Former-commit-id: 556aa155a04f3390d3b961372aa7fa365d77b5da