Fix deadlock in BaseDataProducer.
Move calls to abstract method outside the synchronized block. When they
are in the synchronized block it can cause a deadlock in the following
way. If two classes have locks and they interact through callbacks the
aquiring the locks can have a mixed order.  The mixed order causes a
deadlock.

Bug: 276436535
Test: Run foldable samples.
Change-Id: Ie71ec56e7ba43976fee3e69f74ff386c79c96a7a

Former-commit-id: e214c94a8aa3a386c41eb0238a68ce5b00129b57