Store desktop tasks per display id
Store active and visible tasks per display id.
When notifying the listener of changes, include the display id.

Adding support to controllers to bring only tasks to front from current
display when switching a task to freeform.

Using tasks display id when deciding if a fullscreen task should switch
to freeform on launch or if freeform should switch to fullscreen.

TODO:
- add display id to show desktop calls
- update sysui state when tasks are shown/hidden on other displays
  (currently only default display is supported)
- tests for desktop mode proto 1

Bug: 278084491
Test: atest DesktopModeTaskRepositoryTest DesktopTasksControllerTest DesktopModeControllerTest
Change-Id: I6ebdc4c9b444c22bf7fe58ce82f8db6544fac5e5

Former-commit-id: 2ee5002ae5c9004787f2565bd5b6049368bad40c