fix(#MagnificationThumbnail): thumbnail magnifying area position is not correct sometimes
In MagnificationThumbnail, mThumbnailView size is 0 when the layout has not added to windowManager ever, and thus the magnifying area position x/y calculations would not be correct. Therefore, we use the magnification bounds to calculate the estimated size if the size is 0. If the mThumbnailView is added to window and its size is update properly, we can use the real size to calculate the magfying area x/y after.

Bug: 278509054
Test: manually
      atest MagnificationThumbnailTest
Change-Id: Ifdc84b178e264694061b3b15cbb496ab8f0374b3

Former-commit-id: eaab4f2517c48157bdeca328ee49af80c39c9025