Collecting  Request:Count metrics
This sets up the collection of the request:count metrics, with unique,
unknown request strings, cut off at a certain point (20 characters to
the end). This is in line with our designed plans, but the exact cut off
point may be discussed. The idea is that the server side context is all
that's needed for queries.

Bug: 271135048
Test: Build + E2E Test in progress (won't submit till that is done)
Change-Id: I5420f6dd09f5b52f71926ad20d89d82f9b369b24

Former-commit-id: dc02288f2316bd700c7126db20577213a77813d5