Delay clipboard entrance animation until image loaded
When an image is copied, we load the image in the background but
start the entrance animation immediately. If it takes non-negligible
time for the image to load, this effectively means we animate in a
"blank" preview, then instantly replace it with the real image once
it's loaded, which looks jarring.

Instead, hold the animation until the image is loaded, with a timeout
(currently 300ms). If we don't load the image within the timeout
period, fall back to showing the "copied" preview (same as we do
if we get an exception while loading the image).

It's now possible for us to be "up" while not actually showing any UI
(during that timeout period), so we also turn off tap-outside-to-dismiss
until the animation starts.

Bug: 277586289
Fix: 277586289
Test: atest
Change-Id: Ie6015f712c8993d234bc2bc657c9a81bf549e6e6

Former-commit-id: 7f235677796d44d1d795404379a314cc48b3eb28