Allow fixed rotation for transient launch with visible wallpaper
This aligns the same condition as legacy transition. Otherwise
the swipe up gesture will be interrupted by display change.

Also link fixed rotation transform from wallpaper target if wallpaper
is visible. Previously if it will be ignored if visibility is not
changed. That may cause wallpaper to show in wrong orientation when
running swipe-up animation. The legacy code works normally for the
case so just reuse it.

Bug: 278615635
Test: Swipe up an app with FLAG_SHOW_WALLPAPER to home.
Change-Id: I14e468224457205c7268508d6051037877760a4d

Former-commit-id: 112d53bc99833b15ca7b392588a59472a4651551