Clear IMM#mCurRootView when needed
This clears the mCurRootView property from InputMethodManager
when it becomes inactive or loses window focus.

Before, when going from an app that had the IME shown straight
to the launcher home screen, and tapping on the Google search bar,
this would attempt to show the IME but would sometimes fail. This
is because the show flow would progress further than normal, due to
the leftover mCurRootView value in the launcher process.

Clearing this value ensures the erronous flow is cancelled early, and
due to a subsequent retry, can always succeed in showing the IME.

Test: atest android.view.inputmethod.cts.FocusHandlingTest#testClearCurRootViewWhenDifferentProcessBecomesActive
Bug: 276742733
Change-Id: Id08dece0a6b7531da6d6f8c456cd2e1d6489a9be

Former-commit-id: a981617a4cba56b42e804e442e7739afb7c823ff