ignore call streaming msgs if the adapter is destroyed/not set
The D2DI team hit an NPE while testing out the call transfer/streaming
stuff. The NPE arises in StreamingCall whenever a call streaming changed
msg comes in and the StreamingCallAdapter is null.

Ideally, we should be wrapping any adapter request with an NPE check and
ignoring messages if the adapter is destroyed or not set.

bug: 277119309
Test: manual
Change-Id: I90a5c1dc16dfda9f3c92316b2f4f1655166a59b9

Former-commit-id: 6e377858ffbce87b84d74322cb6e6a710ef26ab8