Stop seekbar listening when carousel is not visible
The seekbar would continue to update if the media was swiped away while
still active, even though the QQS and lockscreen host locations would no
longer display the carousel. This updates the visibility status to
account for whether there is actually any "active" media when the
carousel is in those locations.

Bug: 278145877
Test: atest MediaHierarchyManagerTest MediaCarouselControllerTest
Test: manual - swipe away while playing, seekbar only updates in QS
Change-Id: I0e1bd89ec575f239d8fdb09b5d1289b9d184c1db

Former-commit-id: 07fd8cc1020350e811ed7a6773803919494e804b