Fix flicker of QS footer actions when restarting SystemUI (1/2)
This CL fixes a flicker that could happen with the footer actions View
when SystemUI is starting. This CL is changing the default visibility of
that View to INVISIBLE instead of VISIBLE.

I also had to make FooterActionsViewBinder a class instead of a static
object to make it mockable in QSFragmentTest.

Bug: 263699931
Test: atest FooterActionsViewModelTest
Test: Manual, crash SystemUI
Change-Id: I0f4344cf52dcb00a8b5b38d8a259c0188f13433d

Former-commit-id: 3b9787579c77098522f39e350b948aab5a5f6602