Simplify arrow stretch width interpolator and edge stickiness
Previously the arrow width was based on the difference from
the reactivation point and fully stretched width point.
This caused some discrepancies when these two values were approaching
equality. In this scenario, the width animation would differ
based on the reactivation point which is undesirable.

This change removes this invisible threshold and uses a displacement
based method to calculate width.

Bug: 278083360
Test: Manual
Change-Id: If992c829dc285b883951ee4852cec89e1d707d9f

Former-commit-id: a63e464ae76d2c62d1e684b8986683af9c90ed2d