Fix stage stay empty if close transition request before task vanished
The close transition will not contain trigger task if there are multiple
activities in the task, and only finish the top activity. So the
StageCooridinator won't prepare pending transition for it.
But the whole activities might be finished during collecting, so when
transition ready, the task should be closed and dismiss split screen
if the task is the last child in the stage.
When above case happen, prepare the dismiss transition at
startAnimation.

Bug: 278060807
Test: manual, Launch app info, entering split with another task, then
press back to dismiss app info. Verify another task should become
fullscreen.
Test: manual, entering split, create a new task in one of stage, verify
nothing break.

Change-Id: I61c23511723d93f7a99eb2b2b08ab0b7dc118dcb

Former-commit-id: 0e5ebc1c7bfafecd9e263e8107306c659808c3b4