Update current device when volume control id changes
The volume control id determines the routing session info
associated to the media session. This means when transferring
remote to remote, the check before this changewon't detect the
playback transfer, as it only checks for REMOTE <-> LOCAL.

With this check, we will update the current media device whenever
the media session advertises its association with a different
routing session info.

Bug: 269985505
Test: atest MediaDeviceManagerTest
Change-Id: Ifdbec9e6353602fd9e1774f3aa8d97363f8e506d

Former-commit-id: bf91f0b8ce0ebb4c597cae7d8da0fc6fd4323cce