Check for nullness in WallpaperService onDestroy
There are cases where the WallpaperService API is misused and an instance is created manually. In those cases, onCreate is not called as expected, but onDestroy is.
As mBackgroundThread was being created `onCreate`, it ended up being null in onDestroy.

Adding a null check to avoid breaking those places.

Bug: 275361339
Test: androidx.wear.watchface.editor.EditorSessionTest#commitWithPreviewImage
Change-Id: I89076da179493fa282f3e59895cfa4f4bfb6bdfc
Merged-In: I89076da179493fa282f3e59895cfa4f4bfb6bdfc

Former-commit-id: 784a9ec804b9733e60f45f9afcd55a79ed9bca8f