Pass profile userId to notifyLockedProfile
When the stack originates in the primary user and then a profile
activity is added on top, RunningTaskInfo.userId will reflect the
primary user, but this is not the user for which WorkLockActivity should
be launched. With this CL profile userId is passed along the
RunningTaskInfo into onTaskProfileLocked, so that WorkLockActivity is
started for the right user.

Bug: 279024529
Test: atest WmTests:RootWindowContainerTests SystemUITests WMShellUnitTests WMShellTest
Change-Id: I7755243edf73546dc840d0107d0a6bdb695eb45d

Former-commit-id: af06141f18d54f8eb48f0a59447bee7ab1930cf3