Don't crash system_server due to bad event time stamps.
Right now when we get bad event times in PowerManagerService we throw an
exception which brings down system_server. This is useful because it's a
strong signal that something has gone wrong in the system, but brings
down stability metrics until we can fix the root cause. Instead, just
lot a WTF -- these also got logged server-side for us to investigate
issues.

Fixes: 269510424
Test: atest PowerManagerServiceTest
Change-Id: I1e6cd2d4e1f869fbd5885434b63c007653953cb3

Former-commit-id: 8dfd988e2576554ed7ed49a27618cc60cbe56bd3