Let gesture detector take display ID as param (2/n)
The gesture detector was taking the DisplayTracker and always use the
default display to system UI as the display to monitor input events.
To support the effort to handle transient animation when the bars are
hidden by the app, we need to support a multi-display awared
environment.

This change simply altered the constructor params and make corresponding
changes to its caller to make it be possible to monitor inputs on any of
the given displays.

No actual funtional changes in this patch.

Bug: 277290737
Test: m
Change-Id: I05ea38f409e3f2e7282b608dd841855e80559f3c

Former-commit-id: 0a141fd6fa614024c6d6d3deb8bf8b321eb1f597