CSD: Fix initial logic for mSafeMediaVolumeState
When refactoring the safe hearing logic into the SoundDoseHelper we
simplified the msg enum logic for configuring the safe hearing state.
This lead to an issue allowing the forced init messages to be removed by
the config changed message.

Reintroduced the different messages for state configuration which should
fix the issue of having devices with MCC 0 not configured for safe
hearing.

Test: check state in dumpsys
Bug: 278822191
Change-Id: I515ca3e5dec3f0d1facf752e6f82886b9760bb04

Former-commit-id: 5e7759ba5e699243cc2362edc8fd91ad6c6cf985