[Unfold animation] Move unfold animation background under wallpaper
We add a color layer during unfold animation so
when the wallpaper is not visible (e.g. when an app is opened)
we will show our colored background instead of black pixels.

But when the wallpaper is visible the background is shown
on top of the wallpaper which is undesirable.

Removed the parent of the background layer so it is attached
to the root of the surface flinger hiearchy and below the
wallpaper layer.

Bug: 272703608
Test: manually test unfolding, folding when on launcher,
 a fullscreen app and split screen tasks are visible =>
 checked that we show the background when wallpaper is
 not visible and do not show it when wallpaper is visible
Test: attached winscope traces
Change-Id: Ic7f7ad855055b5de87db912a7dbaf3112bcf289c

Former-commit-id: cde805e99517873d7d7a6a25cdea099055573078