[flexiglass] Container config - domain layer.
The container configuration portion of the Flexiglass framework allows
OEMs to configure one or more Flexiglass containers with scenes. Each scene is a distinct bit of UI content that the
user can navigate to (for example by swiping in a direction) or that can
be navigated to automatically through the SceneInteractor methods.

Bug: 279501596
Test: Included unit tests. Manually tested through a testbed app that
uses Jetpack Compose to implement a simple UI to traverse all scenes.

Change-Id: I28708a55963ae79d48e91f2dcef06eeadea167c4

Former-commit-id: 898bcd5c914ae8c64bdcf1be2fc872134e19c708