MediaCodecInfo: consider only critical flags for level support check
isFormatSupported() checks that the format confirms to the highest
level advertised by the codec for the desired profile if the format
contains a profile.

This additional check uses a capabilities object synthesized from
only the profile/level, so consider only the format keys that
are relevant for a level check for this.

Bug: 278575882
Change-Id: Ide3782882216135e5fac4f53bab21123a73f87b3

Former-commit-id: 7c887bf67c48602c90eaf6267530cd863efcde0d