Verify enableOnBackInvokedCallback at the app-level if ActivityInfo is null
The `ActivityInfo` should not be `null`, otherwise we will not be able to check if the activity has the `enableOnBackInvokedCallback="true|false"` flag.

Unfortunately, there is no good default value in this case.

Our options are:
- Use the application flag, but this may cause random behavior that is difficult to troubleshoot.
- Assume that the activity has `enableOnBackInvokedCallback="false"` in the manifest, which might have strange behavior but at least it would be a "known issue" (for example, if you activate predictive back and it doesn't work in that activity, you can check that XYZ is correct).

Regardless of the reason, I would add a warning or error message explaining the problem.

Test: mp droid
Bug: 271860402
Change-Id: Ie91376ce18eb0a8d5528339a68e4db88a57cff86

Former-commit-id: 04cb994830dd5ce917b611f49aa374e3526189a9