Unload the duplicate unused managers when they conflict
Duplicate versions of several clocks are being included in the
build and loaded by the plugin manager. This prevents those
duplicates, which are already ignored, from continuing to consume
the majority of their memory. I'll need to follow up to understand
why they're being included and loaded in the first place.

Bug: 270860591
Bug: 278073250
Test: Checked device profile
Change-Id: Ia642f640b6bec382bf672d4437f56b5424e9347d

Former-commit-id: b898e57082e5950d47ff753c54d0819827dcdccc