Use IBinder to compare callbacks in remote provisioning service
We were comparing callback types, but these are not guaranteed to
be equivalent when a client calls into our service with the same
object.

Instead, call asBinder to get the underlying IBinder, which is
guaranteed to be the same as long as the remote binder is the
same object.

Bug: 275653768
Test: RkpdAppUnitTests RkpdAppIntegrationTests RkpdAppHostTests
Test: RemoteProvisioningServiceTests keystore2_test
Change-Id: I0d9417234ffab1bb1e892bf4a93760175f3f29c5
Merged-In: I0d9417234ffab1bb1e892bf4a93760175f3f29c5

Former-commit-id: c55fb4a8942ff0830807718c54f032d9162f4997