Put activity-start into its own transition when recents collecting
If a touch through gesture-bar starts an activity, the touch itself
immediately starts recents, then the app (receiving the touch-through)
might start an activity. Because activity-launches must be handled
in sync (due to the API requiring a return value), it can't wait for
an ongoing transition like normal. Instead, if there is an ongoing
transition, it usually just gets added into it. Normally this is
fine (expected, even, for trampolines); however, in the recents
case, we don't want this because it causes a flicker.

In this CL, we basically allow startActivity to create its own
transition in parallel as long as the transition it is "interrupting"
is recents. This way the animations for recents and activity launch
are isolated.

Bug: 276755325
Test: in gmail, flip back-and-forth between it's activity-tabs by
      tapping through the gesture area
Change-Id: Id633baf17445938145132711b556effb852a5aea

Former-commit-id: 7f51866c1154b84a25afc9a617dad4730d2c449a