Store the source ID in InsetsFrameProvider
This CL stores the source ID but not the owner, the index, and the type
in InsetsFrameProvider. In this way, we write less info to the parcel.
Also, we won't use the BinderProxy to generate the hash code for the ID,
but will use the Binder object directly. Because the hash code obatined
from the BinderProxy might change.

Fix: 272230840
Fix: 279476217
Test: atest InsetsSourceTest WindowContainerTests
Change-Id: I0c6434ff19c33e4bb939dee76f49ccb3322735d2

Former-commit-id: bed7ef2e396726af0782d5c5dbcb6c5062e0d0b4