Remove RECORD_AUDIO requirement for ST listModules
No reason to require RECORD pre-flight permissions to query dsp module
properties. CAPTURE_AUDIO_HOTWORD, which is a non-runtime permission,
should suffice to ensure only valid ST clients can call this method.

We will still require RECORD for the rest of the methods, which actually
begin to interact with the hardware.

Bug: 272147641
Fixes: 279839210
Test: SoundTriggerManagerTest
Change-Id: Ibf61fe8f26169dff754dbd40f385793b325707de

Former-commit-id: 87a4ad07d2c8d8088e78e8770bae3a3130625d14