Implement InputConnectionWrapper#takeSnapshot()
This is a follow up CL to my previous CL [1], which forgot to
implement

  InputConnectionWrapper#takeSnapshot().

As a result, it has alreays returned null even if the wrapped
InputConnection does implement InputConnection#takeSnapshot().  This
means that InputMethodManager#invalidateInput() had always fallen back
to slow InputMethodManager#restartInput() path if an app returned a
wrapped InputConnection for example.

With this CL, wrapped InputConnection can also start taking advantage
of fast restart-input mechanism unless it's further overridden by the
app developer to return null for some reasons.

 [1]: I8e00dacce5ca915e276fdd8288ffc9167eb01a26
      94e387c9cc7a12c69230e944153346f068c35d87 [formerly dae47962355ebca330f8fe57e3aa48f4d20a573a]

Test: atest CtsInputMethodTestCases:InputConnectionWrapperTest
Fix: 263677142
Change-Id: I9064a4008d688f2e1188c8d78337f0454c101465

Former-commit-id: 2e376ce96c580589873d12f0192a0c7265f42ad3