Introduce bouncer delay behind flag
Whenever the flag is enabled, the 250ms bouncer
delay is enabled as long as face auth is enrolled
and could run or if active unlock (regardless of whether
it's actually available) could've run based on the device
state.

This doesn't yet use the true active unlock available signal yet
since the API isn't available.

Test: atest PrimaryBouncerInteractorTest
Bug: 267322286
Change-Id: I1f87d2e4848f112880dad64995309ed384407b63

Former-commit-id: b585614729023358754d45f1517b07fd2527e3ec