Enable visual feedback when entering FRP pattern
When the original lock pattern is verified for Factory Reset Protection
(FRP), the original setting of LOCK_PATTERN_VISIBLE is unknown.
However, it's likely to be true, since that is the value that gets set
automatically when a pattern is first set.  It would only be false if
the user went into Settings and disabled "Make pattern visible".  So, to
be consistent with this, fix the low-level default to be true instead of
false.  This enables visual feedback when entering the FRP pattern.

Bug: 270013005
Change-Id: If295999da34510c17ebfb323dd5215418eaa0852
Merged-In: If295999da34510c17ebfb323dd5215418eaa0852
(cherry picked from commit 5162566c0fae79d7e6afd7348100a39296d3ec37)

Former-commit-id: e46d3522d0f4b92efc95a9f62a366b04549733ad