DO NOT MERGE Increase notification channel limit.
This was previously reduced because it can cause an out of memory error or be abused to trigger a permanent denial of service, but it breaks some messages apps in Android 11, as they are creating too many channels. Rolling it back until the apps are fixed to reduce user impact.

Bug: 279447569
Bug: 261723753
Test: N/A, this was previously 50k
Change-Id: Id8da382f812d4abb8db723c40a61366a7402da4f

Former-commit-id: de7c43eb607fbc33677c4391586075861950747a