[DO NOT MERGE] Update quickshare intent rather than recreating
Currently, we extract the quickshare intent and re-wrap it as a new
PendingIntent once we get the screenshot URI. This is insecure as
it leads to executing the original with SysUI's permissions, which
the app may not have. This change switches to using Intent.fillin
to add the URI, keeping the original PendingIntent and original
permission set.

Bug: 278720336
Fix: 278720336
Test: manual (to test successful quickshare), atest
SaveImageInBackgroundTaskTest (to verify original pending intent
unchanged)

Change-Id: Icad3d5f939fcfb894e2038948954bc2735dbe326

Former-commit-id: 589ce3909c6e9e30f073df86e7de4503854a032a