keyguardGoingAway wallpaperCommand
Trigger a wallpaper command when
ActivityTaskManagerService#keyguardgoingAway is called.

This command is not sent if the user has chosen "None" in the screen
lock settings, (in this case pressing the power button goes to home
directly).

Otherwise, this command is always sent after exiting the keyguard,
at the moment when the launcher appears. In case there is a bouncer,
this command is sent after the bouncer finishes hiding. If not, the
command is sent right after the user swipes to go home from lock screen.

Bug: 216838820
Test: manual (put a log in onCommand and verify it's properly called)
Change-Id: I0eed455c16bb6ec92a3789d89d7197d20ca15ee1

Former-commit-id: c6778954399d71d3dbb1923c6dad54bad5521f47