Add support to move task to another display
Create an API in DesktopTasksController to move a task to another
display. It will rotate through the existing displays and move the task
to the next one. If the task is already on the last display, it will
move it back to the first.
Display list is ordered by display ids.
If there is only one display, the method call is a no-op.

Adding a dev option to window caption menu that is behind a flag. This
dev option can be used to move the window.

Bug: 278084491
Test: atest DesktopTasksControllerTest
Test: use a virtual device with multiple displays, use select button to
  move an app between displays

Change-Id: If30b48d04dfa86b64b80e39e63e365bf921a6f92

Former-commit-id: 44fdf07f6573d3cbefd52091352ce33f28fda26f