AudioDeviceBroker: Fix not setting proper call volume for LE Headset
When switching from A2DP to SCO device for the IN_CALL mode, the new
device type stream volume gets applied. However in case of LE Audio
devices, the device stays the same, but when the mode changes to IN_CALL,
the LE Headset is not yet set as a communication device so we must do an
additional stream volume level alignment for the communication stream
when it gets activated.

When the call ends, the mode change handler will restore the previous
mode stream volume for the LE Headset device.

Bug: 278197636
Bug: 245638147
Test: tested manually
Change-Id: I08b0052c799a657b31d6e1a1db2ee88573ae5e83

Former-commit-id: 19a459b78dc56de3eae5a7d0a16891e1d8b41b49