Have GrContext hold a strong ref to VulkanManager
VulkanManager is already ref counted and can be reused
across GrContext's thanks to being shared with
HardwareBitmapUploader.

So take advantage of this + the newly added skia API
for GrContext being destroyed to have GrContext hold a
strong ref to VulkanManager which in turn keeps the
VkDevice alive. If somehow we never recover from this
we'll just keep using the initial device when creating
the next Grcontext which should be fine. Memory usage
will be sub-optimal but that's better than crashing
probably.

Bug: 266626090
Test: unable to repro bug issue, so speculative fix + it didn't get
worse at least?

Change-Id: I2038dacea679c29fd098462d4cf9d298e1215fb2

Former-commit-id: 9fc3d279aee8175e0875f0cbe13d7cdfe92bbfa3