Forward auto-enter property to app's onPause
Pause happens in a different code-path in shell vs
legacy during auto-pip. This path doesn't have auto-enter
context so the app was being told to pause without
being told the pause was due to auto-pip. This CL
adds a state to the ActivityRecord so that when it
pauses, it can send the corresponding autoPip value.

Bug: 280065861
Test: play video in netflix, swipe-to-pip.
      Or, monitor logs and check that pause during autopip has
      autopip set.
Change-Id: I37ef091f3daf483cd12b5c20a796170f49f6db08

Former-commit-id: 07cac275a676ef6df8de92f7f27a69274f3ca89e