Return right activities in getTopVisibleActivities method
Currently, it will only return a top visible activity for every
visible root task in the getTopVisibleActivities method.
On a split screen, there will be two visible activities in the
same root task. It will miss to return an activity when using
the getTopVisibleActivities method on the split screen.

We need to return right activities on the split screen in
getTopVisibleActivities method.

Bug: 245507506
Test: atest CtsVoiceInteractionTestCases
Test: manual test
Change-Id: Id9df59be04bfe805a160d41f044386c3d20eecb6

Former-commit-id: e8e9069b3a93b430bd8bd144ed8f792fa810fe76