Do not show Keyguard Presentation when occluded
..., in concurrent mode, and if it is the display being used for
concurrent mode.

When windows with FLAG_SHOW_WHEN_LOCKED are showing on top of keyguard,
the keyguard state becomes "occluded". When this happens, we should
not show the Keyguard Presentation on the display that's being used
for concurrent mode. This fix allows activities with
FLAG_SHOW_WHEN_LOCKED to present content to the non-default display
when keyguard is locked.

Fixes: 270671994
Bug: 271317597
Test: manual (b/270671994 comment#10)
Test: atest KeyguardDisplayManagerTest
Change-Id: I79235fdd42ee23f2f4178397937eec859c56d844

Former-commit-id: 1de9f064562e47323fee8fc1537e0f419bd48498