Setting up the emit of the aggregate phase
This finds a proper spot to emit the aggregate phase of track 1. Since
we emit final phases of both track 1 and 2 using the same information
right afterwards, the aggregate phase can go right before that. That's
because it needs to fully collect the authentication entry clicks, which
may loop quite a while after the candidate finishes.

Bug: 271135048
Test: Build and won't submit without E2E Test
Change-Id: I94e95ead6e558fcb1f6be10d617f311b65eed7c3

Former-commit-id: e68cc2f42b3653f7f513e8e53f35b39b81215318