Make rotation leash for rotated multi-activity pip expand
When expanding a multi-activity pip, the activity gets reparented
back to its original task. For expanding into a different rotation,
the Task is what held the new rotation, so when the activity is
reparented, it loses that transform. In this case, we need to
create a rotation leash to provide the transform in the task's
place.

Bug: 277452732
Test: play video in netflix (landscape). swipe-to-pip (portrait),
      expand pip (back to landscape).
Change-Id: I70318e2808e46158d20a721e8595651ba3ac8460

Former-commit-id: a35b10dd0d7ba6e80d1a04996b1df3ddea66db59