Cancellable gesture should always have cancellation token
With the current CancellableHandwritingGesture,
cancellationToken is only attached at the time of writeToParcel. However
if editor is in same process as IME (i.e. IME contains the editor),
writeToParcel will never be called and token will never be set.

This change refactors code to not unbeam if we didn't beam cancellation Token.

Bug: 280268494
Test: atest InputConnectionEndToEndTest InsertModeGestureTest
Change-Id: Idb9292fa087d98395d778dd3cd4a528554909ffc

Former-commit-id: 59845d242f52708e5ae3cd8123b5bb5b81683665