Warn the user when a streamed app requests authentication
When a user is streaming an app running on a VirtualDisplay owned by a
VirtualDevice, and that app requests authentication via the
BiometricPrompt api, the streamed view of the app appears frozen and
the user may not understand that they need to go to their phone to
complete the authentication. This CL adds a toast we show in this
situation to help the user know what to do.

Bug: 246425400
Test: atest AuthServiceTest VirtualDeviceManagerServiceMockingTest
Change-Id: I5c88832c948545e9aab1eb23cca97a3165008dae

Former-commit-id: 734a40cf8f67542e3ee9fdcab942416caf5d761e