Modify MANAGE_ACCESSIBILITY to accept roles
This is the first step in allowing Exo and the Chromebook to act
on streamed screen content on behalf of its accessibility users.

It was assumed Exo would have access to this permission, but since
it's not signed by the platform, the signature-level permission is
not automatically granted. This fix allows only pre-installed,
privileged apps that have this role, and which have asked the user
for consent to stream and communicate with apps on the mobile device.

Bug: 280661435
Test: atest AccessibilityDisplayProxyTest, PermissionPolicyTest,
CtsCompanionDeviceManagerUiAutomationTestCases, CtsRoleTestCases

Change-Id: I5c0d4d0a44344a25c41e944bdfbdfaf696da68a3

Former-commit-id: c001f552b4f3b34826f6d8a427074d5f1626e9a2