Add @SysUISingleton to repository switcher classes
The Mobile and WiFi repository switcher classes never got the annotation
to be created as a SysUISingleton. This is probably a mistake caused
by the way we had been providing them; By using a @Binds method on the
StatusBarPipelineModule, it was not immediately obvious that the classes
were not singletons. But believe me, they should be singletons.

Test: manually determined that there is only 1 Mobile and Wifi switcher
instantiated
Test: CarrierTextManagerLog gets current info when airplane mode
switches
Test: manually verified that the "Wi-Fi Calling" text shows when in
airplane mode
Fixes: 278254374
Change-Id: Ifdb5c4b2977020f6ab349b97cfe286c7acbedaa9

Change-Id: If2d87039042fb0a2d3bec42efb1d2d45fa015c04

Former-commit-id: 27617f810f675ab351869bbcbbfeb881be2b35fd