Do not create window decorations for split roots
Split roots (either the single-top root or stage roots) were usually
excluded from having window decorations by the exclusion logic for
non-leafs that exists in FreeformTaskTransitionObserver. However,
once split is dismissed and the child tasks are reparented out of the
stages, the two stage roots become leafs and if they were included as
part of a transition, a window decoration was being created for each of
them. This change excludes split roots from having a window decoration.

Bug: 280499263
Test: enter split, dismiss split by dragging the divider, observe
there's no crash

Change-Id: I4d563176a75512232fc853232b3666d511e4c922

Former-commit-id: 1bd14bc4e399ee6439b638f7bc0077879da38899