[SB Refactor] Use the calculated mobileIsDefault for failed connection.
This is not directly related to b/225902574 but I noticed it while
looking at that issue.

In the case where carrier merged comes through a WIFI connection, it
would never show the ! because mobileIsDefault=false. However, the
carrier merged connection could be non-validated, in which case we
*should* show the ! (but wouldn't right now).

This CL updates the `isDefaultConnectionFailed` calculation to also
show the ! if carrier merged is default and not validated. It does this
by using the interactor's interally calculated `mobileIsDefault` flow,
which includes carrier merged (as opposed to the repo's
`mobileIsDefault` flow, which does not include carrier merged).

Bug: 225902574
Test: atest MobileIconsInteractorTest
Change-Id: Ic79cc1755a67400de067a51abcf654b1294b7fe1

Former-commit-id: 970a8f60258a2ae2c8f4b9528a37808dda45fcc1