Move directional spacing logic out of ViewEntry.
This changelist delegates calculating the spacing between complications
to the DirectionGroup rather than the ViewEntry itself. By calculating
the margin up the hierarchy, more conditions can be accounted for
without passing the information down to the ViewEntries.

This changelist also renames complication margin to directional spacing
to better reflect how the value is used.

Test: atest ComplicationLayoutParamsTest
Bug: 280595239
Change-Id: I0a83ce98bcdd42d1100ebfe4ef2ab49bb2b95dc3

Former-commit-id: 379ef492ce857ed41e0425fb10ba29387aa8f8c1