SilkFX: fix gainmap metadata editor size
Make the gainmap metadata editor more robust against different display
sizes and settings by changing the main layout to "wrap_content" instead
of using a fixed size.

This ensures that things should still work reasonably well even if your
display setup is significantly different from what I tested with (eg.
smaller display size, or larger font size). In particular, it should
help avoid the buttons at the bottom of the editor being cutoff and
making it so that you can't dismiss the editor.

Note that one could still run into issues with extreme situations (eg.
if the display is very very small or the font size is very very very
large), but this should take care of 98% of the edge cases that
previously would be problematic.

Bug: 280889052
Test: verified display with Font Size increased
Change-Id: I4612625ac29d6227c2852b50902098aad43f3d22

Former-commit-id: 792ebab631a427998b4abe6fcbd383d5b9149900