Update the Aod Icon colors after the HUN->AOD transition
The NotificationAreaController is responsible for adding the
StatusBarIconViews to their appropriate container and update their
looks. An icon is usually added through updateNotificationIcons(List<ListEntry> entries),
which adds the icons and updates their colors. It works well if AOD icons are added through this path.
However there is another path through onFullyHiddenChanged(boolean
fullyHidden) and if an AOD icon becomes visible during this update, we
miss to update its color. In case of black PNG icons it resulted in an
_invisible_ notification on the AOD.

To trigger this path:
 - show aod
 - turn off faceauth bypass
 - post a Pulsing notif with a black PNG icon

Fixes: 276353191
Test: post notifs on AOD and observe their color
Change-Id: I0c1e97e9862d8d100e8710f6d825200603f0b2a4

Former-commit-id: bd114eb6fbac193994a13b473ee68d388b8b3999