Add a comprehensive test for Notification.visitUris()
This test uses reflection to find every available setter method on objects and builders and calls them (generating parameters recursively where needed). Thus it's a bit slow and possibly brittle (could be broken by new methods anywhere in the object graph that it doesn't figure out how to handle), but it should be comprehensive.

Fixes: 279441526
Test: atest NotificationVisitUrisTest :)
Change-Id: I77036284d497f6eab705d33f743c6aaad7da4d31

Former-commit-id: 9d86c08bf92f9892616be9e84f87948b0a0b520e