Fixing broken shade after unfolding during expanding QS
The goal of this change is to transmit `QuickSettingsController#computeExpansionFraction() > 0` to CentralSurfacesImpl so in this case we can also close QS before turning screen off.
Now CentralSurfacesImpl checks for shade being fully open or shade being in process of opening/collapsing

In the meantime I noticed mIsExpanding really means mIsExpandingOrCollapsing and renamed it like so - that's why so many files are modified.
Also changed isExpanding() function to be more comprehensive and properly handle shade opening on lockscreen.

Fixes: 277909752
Test: NotificationPanelViewControllerTest#isExpandingOrCollapsing_returnsTrue_whenQsLockscreenDragInProgress
Test: CentralSurfacesImplTest#deviceStateChange_unfolded_shadeExpanding_onKeyguard_closesQS
Test: CentralSurfacesImplTest#deviceStateChange_unfolded_shadeExpanded_onKeyguard_closesQS
Change-Id: I22373e6aaf472a9df50c31fb38db26147b4d3d31

Former-commit-id: bffcecfa07368f2886c6293ca2bda18bb48be3ec