Put change type below closing type for a close transition
Assume a translucent task is closing, the next visible task
may be added to transition participant by collectOrderChanges
as the task is moving to the foreground. Then the transition
info will be {CLOSE, top:CHANGE, bottom:CLOSE}.
So the z-order will be:
 numChanges=2
 zSplitLine=3
 CLOSE=3+2-1=4
 CHANGE=3+2-0=5

Then the CLOSE is occluded by the CHANGE incorrectly.

Bug: 281010752
Test: Launch a translucent activity from launcher and it calls
      Activity#finishAndRemoveTask(). The activity should not
      be covered by the icon on launcher during animation.
Change-Id: I1236946c867abf1b93fca39c4dd9ed114118c051

Former-commit-id: 77e889f27177a9d4ed856710499ed9e5da3f16e8