Add foreground check for controlling or requesting device state
In addition to checking if the app is on top, we should also
check if the app is in foreground. An app may be on top, but
can be hidden by keyguard.

Also note that we are doing both checks, becuase for example
in multi-window mode, multiple apps may be in the foreground.
However, only one of those is considered "top".

Bug: 270671994
Test: android.server.wm.jetpack.area.ExtensionRearDisplayPresentationKeyguardTest#testStartRearDisplayPresentation_whenKeyguardLocked_withoutShowWhenLocked
Change-Id: Iae4d67e55612e9df895a494bb1b891b546c262b9

Former-commit-id: 50f9e8baef9639f708a4bb1741c1698c6bafbb92