Fix unexpected starting window removal timeout when in fixed-rotation
Not aware in DC#mayImeShowOnLaunchingActivity has an issue that
it uses findMainWindow() to check its softInputMode visiblity state
but always got 0 because the returned main window is the starting window
vs. the app's base application window, which is not expected.

This problematic logic leads to when CL[1] merged, the starting window
removal mode and timout always set to DEFER_MODE_ROTATION (3 secs)
vs. DEFER_MODE_NONE (100ms) if the warm launching activity's softInputMode
is hidden during fixed-rotation.

Ensure finding the main window with excluding the starting window in
DC#mayImeShowOnLaunchingActivity to fix this case.

[1]: Ie476e89a57f2f64d4d66e722fedeeb1719d9de55

Bug: 268627602
Test: atest FlickerTests:\
      ShowImeOnAppStartWhenLaunchingAppFromFixedOrientationTest
Test: atest DisplayContentTests#\
      testMayImeShowOnLaunchingActivity_negativeWhenSoftInputModeHidden
Change-Id: I3fa33daf20734b673fd89328dca6a8fd432f2969

Former-commit-id: 392226b61e33c6cf85eda51a4e7bb1322f2dab84