Allow rolling back Turn Off Work 2.0
To facilitate it, previous effective value is stored in
DevicePolicyData, with "true" default value to avoid
storing it. When the system boots, if the current flag
value is different from the stored one, suspension is
updated.
An upgrade step sets the stored state to "false" once,
so that devices upgrading from pre-U get their work
apps suspended if in quiet mode.

Bug: 265683382
Test: Manual, via device_config
Test: atest PolicyVersionUpgraderTest
Change-Id: I711ace4cc2f8da6ce4f85aa0f0cde740dba2ba02

Former-commit-id: a855429993433a007f200465a571bab7364dbff4