[flexiglass] Container config - UI layer.
The container configuration portion of the Flexiglass framework allows
OEMs to configure one or more Flexiglass containers with scenes. Each scene is a distinct bit of UI content that the
user can navigate to (for example by swiping in a direction) or that can
be navigated to automatically through the SceneInteractor methods.

Bug: 279501596
Test: Included unit tests. Manually tested through a testbed app that
uses Jetpack Compose to implement a simple UI to traverse all scenes.

Change-Id: I2c925707b0709e4aaf26bca8f94e5e2b3cf6bbcc

Former-commit-id: 213e2ad4758d3e91bde646267ded95b0d38d7996