hwui: send TextureView hint to SF
So that SF could use this hint when choosing the refresh rate.
SF would only try to heuristically calculate the frame rate of a layer
when TextureView is updating. This fixes a bug where SF tries to
heuristically calculate the frame rate for UI animations but fails
due to long frames.

Bug: 280249265
Test: Playing a video on Facebook and observe refresh rate
Test: go/cb-pcmark

Change-Id: I0d54d62b97ff48583fbe3cc0da188fe85810fd5e

Former-commit-id: cab4afeb64bbc6280990b34bd55270b19218ac04