Inject a handler for Keyguard transitions
Keyguard transitions absolutely must not fail because Keyguard, not
being a window itself (more like a mode for SystemUI views), uses the
signal from the transition system to control its visibility.

We're putting a KeyguardTransitionHandler at the top of the list of
handlers so that it has the first opportunity to run anything related
to occlusion or keyguard-going-away flags.

As of now the KeyguardViewMediator is still using the IRemoteTransition
interface to keep the change small and safe-ish. This can be improved.
For example, there's no real need for both sides to have a special
concept of "occludeByDream".

Test: atest ShellTransitionTests
Bug: 274954192
Change-Id: I54e4ab66840c47e686071cab3c22f4c351dc875e

Former-commit-id: 675402855fef2651168d912356e6748dbd53d881