[flexiglass] Composables for scenes and scene container.
Implements the composable functions for scene placeholders and the scene
container, which itself, is a placeholder. Instead of handling true
swipes and real UI, the placeholders have buttons for triggering a swipe
event and no transition animations between scenes. This entire system is
meant purely for testing the scene switching logic in previous CLs in
the Compose Gallery playground app.

Bug: 279501596
Test: Manually verified all scene transitions using the testbed app.
Change-Id: I4b81988b2ea6ef5388714674718e2f742dd1b3a4

Former-commit-id: 0af3d92e5b661cd3e0ad2e454eb8f8d78a12d176