Fix potential IllegalArgumentException from the auto-select flow.
AutoSelect flow animate the bottom sheet directly into empty sheet
because it will then immediately launch the provider activity. However,
there's a native compose bug that throws IllegalArgumentException when
the sheet is animated from/to an empty sheet. The fix was submitted to
late for us to merge into our U release b/ 281791945 so we need a temporary solution:
for now we'd keep a 1 pixel transparent bar when this could happen; it
won't be noticeable.

Bug: 281791945
Test: manual
Change-Id: I668af5316fa795cf833d0907481373d86bb24ed1

Former-commit-id: 35c3307f7d43591939e7ec3c83c5cb57ec2d95a8