Give Keyguard 2 seconds to finish before SLEEP
When we're entering dream or going back to keyguard, we may generate
a SLEEP transition right after to stop long-running transitions.

This has the drawback that it interferes with keyguard transition
animations. So KeyguardTransitionHandler gets a special allowance to
play for a little longer before we bring the SLEEP hammer down hard.

A longer period exposes a problem that if we generate more transitions
shortly after sleep (eg. screen off/on fast) we can catch some of the
new transitions in the crossfire. This can be avoided by keeping track
of which SLEEP transition generated the sync signal and stopping the
sync as soon as that transition has started playing.

Test: atest ShellTransitionTests
Bug: 274954192
Change-Id: I80f6b3144b75c8c2ea39e5706fe03ec28456edb3

Former-commit-id: 4a7d60052473a89503ed75d247b965d9748208de