Revert^2 "Add @SysUISingleton to repository switcher classes"
The Mobile and WiFi repository switcher classes never got the annotation
to be created as a SysUISingleton. This is probably a mistake caused
by the way we had been providing them; By using a @Binds method on the
StatusBarPipelineModule, it was not immediately obvious that the classes
were not singletons. But believe me, they should be singletons.

Test: manually determined that there is only 1 Mobile and Wifi switcher
instantiated
Test: CarrierTextManagerLog gets current info when airplane mode
switches
Test: manually verified that the "Wi-Fi Calling" text shows when in
airplane mode
Fixes: 278254374

Change-Id: Idd766a5378d8deebe50bafe1d564635187ab30ee

Former-commit-id: fefb9a945a9e00e0cc441416a0aa5722e2cdeb94