[pm] shell command to wait for handler(s)
Adding a shell command to wait for the Package Manager handler to finish
handling all messages that are currently waiting in the message queue.
This gives a better indicator for tests to know when to wait for
incoming broadcasts.

Added a similar command for the background handler as well.

BUG: 280604852
Test: atest android.content.pm.cts.PackageManagerShellCommandMultiUserTest
Change-Id: I43c281a4f17edc3982a7885a39b1182be57c4504
Merged-In: I43c281a4f17edc3982a7885a39b1182be57c4504

Former-commit-id: a2d63f09b76a480e34004596db2ae229f8a1cff9