Use ComponentCallbacks instead of ConfigurationChangeListener
ConfigurationChangeListener can be out of sync with theme so
instead use ComponentCallbacks listener. For this to work, it needs
to be registered with a window context.

Updates tests to use the window context bubble controller creates.
Adds a new test to ensure component callback is added / removed
appropriately.

Updates SysuiTestableContext to create a window context, this
ensures that registered recievers for bubbles get tracked.

Test: atest BubblesTest
Test: manual - have a bubble, expand it, change the theme, check
               that the manage button & contents is in correct theme
               along with the overflow button & contents and flyout
             - repeat above with font size, display size, density,
               and RTL and verify bubble UI elements update for those
               changes
Bug: 281748524
Change-Id: Ibdcb680e64bbe81af72ec04318f091941da5fe89

Former-commit-id: 2c4955bcf664e9ed5b9fa5b7fb201c923f9a9b68