Prevent media controls with no title
Apps targeting U+ will be required to include a non-empty title when posting
media controls. If the title is blank, the notification will be
cancelled and report an error back to the app.

For apps not yet targeting U, a placeholder string with the app name will be
used for the title instead.

Bug: 274775190
Test: atest MediaDataManagerTest
Test: manual with test app and YTM
Change-Id: I2db9edd269c2b62ee7de6714875fa098567510b6

Former-commit-id: e3a971d814f33485277dd7b25e2d21e63ca6c334