Retrieve BackupAgent logs after onRestoreFinished completes execution
This fixes a race condition that was introduced in ag/20312251. The logs
were requested right after doRestoreFinished was called. But because
that method is async, this resulted in a timeout while waiting to
retrieve logs if the BackupAgent took longer than 500ms to execute
onRestoreFinished-- which WallpaperBackupAgent will almost always do (it
sets the restored wallpapers in that method).

Bug: 282013010
Test: this class is not unit testable. Tested manually by adding debug
logs and doing a restore. Observed that the wallpaper restore took
longer than 500ms and there was no timeout. The logs were successfully
retrieved after the restore was finished.

Change-Id: I697e1279b9e366da15a6e31a9730e91a8b3d0d0b

Former-commit-id: 2865eb7b2170d0690b748ee79e630a838b9139cd