Metric System Cleanup
This fixes any final issues in the metric collection and emission
system, including logs. The fixes primarily focus on ensuring the
timestamps are properly recorded, the aggregations are done properly for
the total canddiate, and that all types and values are recorded or
collected as they should be and emitted when they should be.
Statuses and TODOs are also fixed here, alongside any final logs.

Bug: 271135048
Bug: 270568738
Test: Build, won't submit without E2E Test (No Affect + Validate Metric Emits)
Change-Id: I3637bf4a84596ba72bdab9ef225e4e5c77ab0bb6

Former-commit-id: 7b1399ec6db0053c1496f966aeb7a6465e485b4e