Prevent alpha transitions from clipping out-of-bounds clocks
In order to do this, we bypass the default alpha mechaism used by views,
and call saveLayerAlpha ourselves on several views. This allows us to
set bounds for this area outside of the measured bounds.

Clipping still appears to occur after a user switch. b/282226587 will
track that follow up issue.

This is effectively a revert of Ia73ea00920ca0147e0acc5d2919b457203b7575e

Bug: 280113068
Bug: 281697986
Test: Manually checked several transitions
Change-Id: I9112efab88facd0f32daa5cd7377b8184658c159

Former-commit-id: 93e23c58dc0c7dd997f20734dafdf37802854f2c