Limit the background FGS start with foreground only permission
Since Android R, background-started FGS can't access the mic/camera
even if the user has granted "while-in-use" permission, so does for
the foreground location permission. Now, if the app tries to start
a FGS from the background with the type mic/camera/location,
the start would be denied, since they can't access those sensors
from that FGS anyway, it makes no sense to start such a FGS.
If the app has the background location permission and tries to
start FGS from the background, this is still allowed.

Bug: 280936347
Test: atest CtsAppFgsTestCases
Change-Id: I56865b4f9e9c317c968a6cbfced5b3b4cdccda79

Former-commit-id: efc4125964bd58542e231e956f86d475fa2f6d65