Replacing overlapping flag value
FLAG_IGNORE_EPHEMERAL shared its value with FLAG_ACTIVITY_REQUIRE_DEFAULT
which was problematic since these flags were used or checked together
in a few cases. A short term solution was to replace the value of
FLAG_IGNORE_EPHEMERAL with another flag value - FLAG_RECEIVER_OFFLOAD -
which isnot used with the former flag.
This will be revisited in the future, with a concrete solution, since
there are no 32bit flags left in Intent.java.

Bug: b/269674648
Test: None
Change-Id: I9e3d282bb12a630d0a170871fe8e2da5d259ad57

Former-commit-id: 2bb931237e5f5fe4b98b3e003851d1a6b6eb4878