Do not set alpha to 1f when bouncer shows
Alpha here was set to 1f everytime the bouncer shows to ensure that the
bouncer shows every time. This has caused issues with the introduction
of a material U background color for the bouncer. We had bugs where the bouncer alpha was 0, which is why alpha is set to 1f here.
Looking at it further, setExpansion is not called in the cases that
expansion is already 0f. This is because we were not setting the panel
expansion to HIDDEN (1f) anytime we hide the bouncer.

Fixes: 281024571
Test: Simpin/Simpuk
Test: password, pattern, pin scrimmed and track
Test: Testing auth and then power off and on and going back to scrimmed
bouncer and tracked bouncer.

Change-Id: I572280daef98da64045e12508a6efd60141a42b6

Former-commit-id: 0e6ba4ffe790a1145fa230c4b795196efcbcff0a