Allow long-press lockscreen popup to be wider.
To reduce the chances of needing to truncate the text in the lockscreen
long-press popup button, places the button z-order "above" the left and
right lockscreen shortcuts.

Please see screenshots at b/279675855#comment16

Fix: 279675855
Test: manually verified that the popup shows properly, in the middle,
between the two lockscreen shortcuts when the font size setting is set
to smallest, default, and largest, on a Pixel 7. Also made sure that
clicking the popup still goes to the lockscreen customization screen,
that the lockscreen shortcuts can still be interacted with, etc.

Change-Id: Ia95a9e11ea5be70960fc45a26271f15c0623b8ad

Former-commit-id: 8f103668a37c37d67cdfea4c90b93bc6fef8859b