Make flushing broadcast loopers optional.
In some cases, it takes a while for the background thread
looper to finish and it seems unnecessary to increase runtime
of existing clients if they don't require this or if there are
no pending broadcasts waiting to be sent in these loopers.
If clients require this behavior, they can always use
the new option to force it.

Bug: 280316424
Test: atest services/tests/mockingservicestests/src/com/android/server/am/BroadcastQueueTest.java
Test: atest services/tests/mockingservicestests/src/com/android/server/am/BroadcastQueueModernImplTest.java
Change-Id: Iee5367dd02ba9e1878fe1d7c0920cf6c0c1c1e2a

Former-commit-id: acbe214d99f689ffb99cab760da7d6bea4ae2735