Individually track vis duration of unseen notifs
Previously, if the user has been on the keyguard for 5 seconds, then we mark all unseen notifications as seen when the device is unlocked, even if some notifications were not present for the entire (or any of the) 5 second visible duration.

We now track for each unseen notification that since it has been posted / the device was locked, that the user has been on the keyguard for 5 seconds. When a notification has reached this threshold, it is marked as "seen on lockscreen", and will be marked as seen for real once the
device is unlocked.

Note that if the notification is updated, the 5 second threshold is reset.

Fixes: 277616032
Test: atest KeyguardCoordinatorTest
Change-Id: I8bcd3e8aef4662970dd52ba5748d7f51ba652ba4

Former-commit-id: c197cbec0d58524241c37eaf152300deb07ccaeb