Add BiometricPrompt implementation with class 3 support.
This adds a new implementation of the prompt behind flag BIOMETRIC_BP_STRONG. The original prompt remains in place temporarily, with minimal modifications, until it can be replaced.

This first phases replaces the top most AuthBiometricView(s) but resues many of the legacy components (icon controllers, panel controller, container view, etc.), which will be replaced and simplified in follow ups. It also includes some changes to framework/sysui APIS required for the class 3 flows that cannot easily be flagged and impacts both implementations.

In addition to class 3 support, the secondary goal of this change is to simplify & address the lifecycle issues with the current implementation, align it with SysUI's architecture guidelines, & enable better unit test coverage (including uidiff).

Bug: 272510026
Bug: 251476085
Test: atest com.android.systemui/.iometrics
Test: manual (use BP integration test app and perform auth with all variations fp-only, face-only, coex)

Change-Id: Ib95906cc7f4837770b778e589ef5e6061e5f39de

Former-commit-id: aa1922acedf22c1d0b3ef288ae39884230e73119