Fix themes of the bubble manage button
Still not entirely sure why this is needed. Both of these views
get reinflated on theme change, and some of their colors are updated
correctly (e.g. font color) but the background color of the drawable
is not.

Pulling the color from the theme attributes and applying it directly
to the background resolves the issue.

Bug:  281748524
Test: manual - have a bubble, change the theme, open the bubble,
               check that the manage menu (and contents of menu)
               render in the correct theme.
Change-Id: I339886d1f1a0ff5b817c106f707518263c67ab07

Former-commit-id: 7a2f2a3a1ecb9d5019fea956a0cd3f8574a6b1f0