Starting sandbox activity should not start a new sandbox process
The sandbox activity is designed to be hosted in a running sandbox
process with a runtime loaded SDK ready to to populate views into it.

If for some reason, the sandbox process is killed while the activity is
being created, default behaviour of the system is to create a new
sandbox process, that would create a process lacking some important
information which cause exceptions, which might cause system crash.

This change avoids creating new sandbox process if the required
information is missing, and it aborts launching the sandbox activity if
there was no sandbox process to host it.

Test: manually by killing the sandbox process while launcing the sandbox
activity
Bug: 278086359

Change-Id: Ibab83a00cdc744e1ff7286de0868f57612e5c5c0

Former-commit-id: 337a29bfe38e3553549cbf2ea7013ae30b733713