[Bluetooth] Consider LE_AUDIO to be an audio-only bluetooth profile.
When we originally added the concept of audio-only profiles in 2019, the
LE_AUDIO profile did not exist. Now that the LE_AUDIO profile exists and
is being used more frequently, we need to include it in our calculation
of audio-only profiles.

This change will mean that, if an LE_AUDIO device is connected but not
active, we will *not* show the bluetooth icon in the status bar.

Fixes: 278982782
Test: atest BluetoothControllerImplTest (verified that the
LE_AUDIO-related tests fail without this change)

Change-Id: Ie58a85d908943ad29bee21b9ccf1a86a34123f27

Former-commit-id: 65f5d0b889bcd01b26af6f71f91fcae04282d133