Relaunching activity to PAUSE state
Relaunching activity locally from ON_START state used to go to
ON_STOP and skipped ON_RESUME. However, the Activity#mDecor
(and the other preserve window handling) is set in
ActivityThread#handleResumeActivity. So, the decor window may
be dropped when relaunching twice.

Relaunch the activity to PAUSE state if it is currently in START
state. This is also the expected state while relaunching the
activity from server side.

Bug: 249451905
Test: verified locally with sample app
Change-Id: Ia24ecc764ff783a1298a706457ab907bf9c52c98

Former-commit-id: f7256f5dd56369b182c084f52f7aa45cfd6a05ca