Adapt Wallpaper backup for lockscreen lwp
Starting from U-QPR1, users can set a live wallpaper on the lock screen
only. The feature is gated through the flag
WallpaperManager.isLockscreenLiveWallpaperEnabled. This CL adapts the
backup and restore logic accordingly when the flag is true. It should
have no effect when the flag is false.

Note: if a post u-qpr backup contains a lockscreen only live wallpaper,
and gets restored to an older device not supporting lockscreen live
wallpapers, the lockscreen wallpaper will be ignored and the home
wallpaper will be applied to home+lock.

Bug: 283091821
Test: backup & restore manually with two live wallpapers
Test: atest WallpaperBackupAgentTest
Change-Id: I98f847c97f42eb65fa626bf601e75daf15d59a45

Former-commit-id: c187b71d9c43eab65fafd6d7042ac59a64696087