Don't show user education if it's hiding
If we get an onWindowInsets change while the user education is
showing (and the user is interacting with it so it's 'going away')
we can end up in a state where we trigger the user education to
show again when it shouldn't.

I've found this pretty tough to repro, I think this is the issue
and should fix it.

Bug: 282618534
Test: manual - adb shell pm clear com.android.systemui
             - adb shell am restart
             - add a bubble, tap / drag slightly on it to show the
               user education
             - tap again, user education goes away
             - repeat this a whole bunch
Change-Id: I05dcfed7466527e26bdce6af5988f5c178203602

Former-commit-id: 2f8ce896dbac5b9d839a5a9874b87bb726deb62f