Improve drag divider dismiss transition.
Before this CL, the split root task will not include in tansition
when drag to dismiss. It leads to users could see flicker when
the top task bounds changed due to no animation on it.

On this CL, it add update bounds too on dismiss transition then
split root could be included in transition. And we could use decor
to do fade-out animation.

Fix: 245472831
Fix: 283045645
Test: manual
Test: pass existing tests
Change-Id: Ie6be85173d9a5b71603aa89858df117be08ea484

Former-commit-id: 9b27cdf8bf474cc45c4910dacf272da2dc228951