Propagate last-ime-shown in the same task explicitly
For example:
 Task [
   top: Activity B with a focused EditText and IME is shown
        Activity A with a focused EditText
 ]
The test expects that after B calls finish, the IME will show
for A automatically.

Previously, the ime-restore relied on a task snapshot when the
activity is finishing. But the snapshot is unnecessary because
there is no task switch. The logic worked accidentally by the
heavy snapshot operation.

Since shell transition should be able to centralize the timing
to take snapshot to reduce duplicated snapshot in a cleaner way,
it is more reasonable to propagate last-ime-shown directly.

Bug: 283413681
Test: atest com.android.inputmethod.stresstest.AutoShowTest \
      #autoShow_backwardWithKeyboardOn
Change-Id: Ic58d1afd6e0e2f7db2136ac60818e3b46f338bc8

Former-commit-id: 3756211c110b83ff4561e4a8adbfe6521638fde2