Override applyLayoutFeatures for ConstraintHelper
This method seems to set the visibility of the subviews based on the
visiblity of constraint helper; however we want to hide some views and
show some views. The method also adjusts the z order of the view, which
seems moot. This will prevent us from updating the visibility of the
enter button and the delete button upon screen rotation.

Fixes: 268589232
Test: Rotate the screen 90, and 180 degrees. Flip open and close the
device. Observer the visibility of the enter and delete button.

Change-Id: Id5d8c01e658aed4b06970b5475228f171848f3c4

Former-commit-id: 13178eddf514208ab206ad43dce45d07592fe0ec