Fix mobile data toggle status not correct
When toggle off the switch and dismiss the confirmation dialog.

This issue is caused by dialog's on dismiss listener not set, only
confirm and cancel listener is set.

Since confirm and cancel will also called dismiss listener, set a
dismiss listener cannot fix this issue.
So when need display confirmation dialog, change the toggle status back
to on, and set to off when confirmed to fix this issue.

Fix: 283259670
Test: Manually with InternetDialog
Change-Id: Ic8f70518a22147f7f01936015d1ac6c341a30993

Former-commit-id: 72117b22aaef4df1c6fe87d6d517b59233f227b1