Reduce PiP flickering when switch apps with orientation change
The fade state of PiP is only known by shell side. So the default
finish transaction of display transition will reset alpha to 1.
That may cause one frame to flicker before the fade animation
animates from alpha 0 to 1.

So for display transition, keep the PiP invisible if it was faded
out (it should be only from fixed rotation). And then after the
display transition is done, fadeExistingPip(true) will be called
to restore the visibility without overlapping with transition.

Bug: 283757841
Test: Make an app enter PiP.
      Launch a landscape app from portrait home.
      And finish the landscape app.
      The PiP should not show one visible frame before
      the fade-in animation runs.
Change-Id: I4f0bd06a29c2537c0e3431a0e5be7e00ae7cd9e7

Former-commit-id: 5103f76ae265f3474a0c9079c5d9748d3467753a