Update nav bar insets when in extractView
Before, when the IME was shown in the extractView (fullscreen), with
gesture navigation enabled, and in landscape, the first time the IME
would show, the IME collapse and switcher buttons would not appear in
the navigation bar. This was because we were receiving 0 insets
initially, but the updated values would not be propragated later on.

Test: open an app where the IME shows in full screen when in landscape
  (e.g. WhatsApp) directly in landscape with gesture navigation
  enabled, show the IME (e.g. tap on search icon), observe whether the
  navigation bar buttons show correctly
Bug: 260053129
Change-Id: I4df22b47a417e3045eee0662bfe73459b266c6a1

Former-commit-id: a17835e7f877c16f69627fb3a661a314d1b5b153